requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-jose[cryptography]
passlib[bcrypt]
python-multipart
python-dotenv
pydantic
pydantic-settings
alembic
pydantic[email]
websockets
python-socketio[asgi]